requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
requests==2.32.3
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.20
pillow==12.0.0
numpy==2.2.6
cairosvg==2.7.1
apscheduler==3.10.4
